+++ cmsplugin_blocks/models/slider.py
@@ -1,4 +1,17 @@
 # -*- coding: utf-8 -*-
+"""
+======
+Slider
+======
+
+A slideshow component which may be similar to Album but with difference that
+a slide item can have HTML content.
+
+Slide items are ordered from their ``order`` field value. Items with a zero
+value for their order will be ordered in an almost arbitrary order (mostly
+depending from item object id).
+
+"""
 from django.conf import settings
 from django.core.validators import FileExtensionValidator
 from django.db import models
@@ -15,7 +28,7 @@ from cmsplugin_blocks.utils import SmartFormatMixin
 
 class Slider(CMSPlugin):
     """
-    Slide container
+    Slide container for items.
     """
     title = models.CharField(
         _("Title"),
@@ -23,6 +36,10 @@ class Slider(CMSPlugin):
         max_length=150,
         default="",
     )
+    """
+    A required title string.
+    """
+
     template = models.CharField(
         _("Template"),
         blank=False,
@@ -31,6 +48,10 @@ class Slider(CMSPlugin):
         default=get_slider_default_template(),
         help_text=_("Used template for content look."),
     )
+    """
+    Template choice from available plugin templates in setting
+    ``BLOCKS_SLIDER_TEMPLATES``. Default to the first choice item.
+    """
 
     def __str__(self):
         return Truncator(strip_tags(self.title)).words(
@@ -45,6 +66,8 @@ class Slider(CMSPlugin):
         See:
 
         http://docs.django-cms.org/en/latest/how_to/custom_plugins.html#for-foreign-key-relations-from-other-objects
+
+        :meta private:
         """
         self.slide_item.all().delete()
 
@@ -60,19 +83,24 @@ class Slider(CMSPlugin):
 
 class SlideItem(SmartFormatMixin, models.Model):
     """
-    Slide item
+    Slide item to include in container.
     """
     slider = models.ForeignKey(
         Slider,
         related_name="slide_item",
         on_delete=models.CASCADE
     )
+
     title = models.CharField(
         _("Title"),
         blank=False,
         max_length=150,
         default="",
     )
+    """
+    Required title string.
+    """
+
     image = models.FileField(
         _("Image"),
         upload_to="blocks/slider/%y/%m",
@@ -86,31 +114,55 @@ class SlideItem(SmartFormatMixin, models.Model):
             ),
         ]
     )
+    """
+    Required image file, limited to enabled image formats from settings
+    ``BLOCKS_ALLOWED_IMAGE_EXTENSIONS``.
+    """
+
     content = models.TextField(
         _(u"Content"),
         blank=True,
         default="",
     )
+    """
+    Optional long text, it will be editable through CKeditor on plugin form.
+    """
+
     order = models.IntegerField(
         _("Order"),
         blank=False,
         default=0
     )
+    """
+    Number for order position in item list.
+    """
+
     link_name = models.CharField(
         _("link name"),
         blank=True,
         max_length=45,
     )
+    """
+    Optional string for link name.
+    """
+
     link_url = models.CharField(
         _("link url"),
         blank=True,
         max_length=255,
     )
+    """
+    Optional string for link URL.
+    """
+
     link_open_blank = models.BooleanField(
         _("open new window"),
         default=False,
         help_text=_("If checked the link will be open in a new window"),
     )
+    """
+    Checkbox to enable opening link URL in a new window/tab.
+    """
 
     def __str__(self):
         return Truncator(strip_tags(self.title)).words(